revisi chat page
