+++ src/pages/GroupPage.jsx
@@ -1,15 +1,13 @@
 import React from "react";
+
 import MainNavbar from "../components/MainNavbar";
-import {withStyles} from "@material-ui/core";
 import ChannelNavbar from "../components/ChannelBar";
+import {GiftedChat} from "react-web-gifted-chat";
+import TopNavbar from "../components/TopNavBar";
+import {withStyles} from "@material-ui/core";
 import Grid from "@material-ui/core/Grid";
 
 import '../assets/styles/chatpages.css';
-import ChatAppBar from "../components/ChatAppBar";
-import GuildDiscovery from "../components/GuildDiscovery";
-import ChatBar2 from "./ChatPage";
-import Container from "@material-ui/core/Container";
-import {GiftedChat} from "react-web-gifted-chat";
 
 const useStyles = () => ({
     root: {
@@ -17,11 +15,10 @@ const useStyles = () => ({
     },
     content: {
         flexGrow: 1,
-        position: 'static'
+        position: 'relative'
     },
     chats: {
         backgroundColor: '#36393F',
-        minHeight: '800px',
     }
 });
 
@@ -35,13 +32,14 @@ class Group extends React.Component {
             messages: [
                 {
                     id: 1,
-                    text: 'Hello developer',
-                    createdAt: new Date(),
+                    text: 'My message',
+                    createdAt: new Date(Date.UTC(2016, 5, 11, 17, 20, 0)),
                     user: {
                         id: 2,
                         name: 'React',
                         avatar: 'https://facebook.github.io/react/img/logo_og.png',
                     },
+                    isAnimated: true
                 },
             ],
         });
@@ -60,29 +58,28 @@ class Group extends React.Component {
                 <div className={classes.root}>
                     <MainNavbar/>
                     <main className={classes.content}>
-                        <Grid container spacing={0}>
-                            <Grid item xs={12} lg={2}>
-                                <ChannelNavbar/>
-                            </Grid>
+                        <Grid container>
+                            <Grid className={classes.chats} item xs={12} lg={12}>
+                                <TopNavbar/>
 
-                            <Grid className={classes.chats} item xs={12} lg={10}>
-                                <ChatAppBar/>
-                                <Container style={{paddingRight: 0}}>
-                                    <Grid container spacing={0}>
-                                        <Grid item xs={8} lg={9}>
-                                            <GiftedChat
-                                                messages={this.state.messages}
-                                                onSend={(messages) => this.onSend(messages)}
-                                                user={{
-                                                    id: 1,
-                                                }}
-                                            />
-                                        </Grid>
-                                        <Grid style={{backgroundColor: '#2F3136'}} item xs={4} lg={3}>
-                                            Member Lists
-                                        </Grid>
+                                <Grid container>
+                                    <Grid style={{backgroundColor: '#2F3136'}} item xs={12} lg={2}>
+                                        <ChannelNavbar/>
                                     </Grid>
-                                </Container>
+
+                                    <Grid item xs={12} lg={8}>
+                                        <GiftedChat
+                                            messages={this.state.messages}
+                                            onSend={(messages) => this.onSend(messages)}
+                                            user={{ id: 1,}}
+                                        />
+                                    </Grid>
+
+                                    <Grid style={{backgroundColor: '#2F3136'}} item xs={12} lg={2}>
+                                        Member Lists
+                                    </Grid>
+                                </Grid>
+
                             </Grid>
                         </Grid>
                     </main>
